Musicribe-py/requirements.txt
pygame-ce
mutagen
pathlib
time
re